FEATURE local dev: Add nodemon
