+++ package.json
@@ -12,7 +12,7 @@
     "start:frontend": "cd frontend && vite",
     "start:local-backend": "run-p start:local-backend:db start:local-backend:gateway",
     "start:local-backend:db": "docker run --rm -p 8000:8000 amazon/dynamodb-local",
-    "start:local-backend:gateway": "ts-node backend/local/main.ts",
+    "start:local-backend:gateway": "nodemon backend/local/main.ts",
     "preview:frontend": "npm run build:frontend && cd frontend && vite preview",
     "build:backend": "run-p build:backend:onconnect build:backend:ondisconnect build:backend:sendmessage",
     "build:backend:onconnect": "cd backend/onconnect && rimraf build && tsc -p .",
@@ -46,6 +46,7 @@
     "husky": "^6.0.0",
     "jest": "^26.6.3",
     "lint-staged": "^10.5.4",
+    "nodemon": "^2.0.12",
     "npm-run-all": "^4.1.5",
     "preact": "^10.5.13",
     "preact-render-to-string": "^5.1.18",
